pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Lanterman <splen1011@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
channels = "^4.0.0"
channels-redis = "^4.0.0"
django-debug-toolbar = "^3.8.1"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
redis = "^4.4.0"
requests = "^2.28.1"
pillow = "^9.3.0"
coverage = "^7.0.1"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
djangochannelsrestframework = "^1.1.0"
django-cors-headers = "^3.13.0"
daphne = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
